requirements.txt
AfricastalkingGateway==2.0
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==1.11.8
django-admin-view-permission==1.4
django-appconf==1.0.2
django-debug-toolbar==1.9.1
django-googledrive-storage==1.3.4
django-import-export==1.0.0
et-xmlfile==1.0.1
google-api-python-client==1.6.5
httplib2==0.10.3
idna==2.6
jdcal==1.3
oauth2client==4.1.2
odfpy==1.3.6
openpyxl==2.5.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-dateutil==2.6.1
python-decouple==3.1
python-http-client==3.0.0
pytz==2018.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
sendgrid==5.3.0
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
